Rework assignments and cases to cloned for each training
Explanation:

There can be scenarios, where an admin wants to update 'assignments' or 'cases'.
This cannot happen when a training is in progress, this may result into unreferenced relations between assignments/questions/answers/cases.
( imagine an assignment, question or answer is removed while a user is interacting with it )
a solution would be to soft-lock the entire assignment structure whenever a 'training' is in progress.
however, IF in the future more admins are assigned with different locations, with overlapping training schemes, there might ALWAYS
be a 'training' in progress, communicating to alter an assignment or case can prove to be challenging.

Thats why is opted to clone all assignments and case data into a dedicated table structure specific for each training.
This means that once a training is started, there is no way to alter the assignments in case of errors/typos.
but does ensure no errors are raised when an assignment/question/case is removed.
this method also makes storing user data easier and cleaner.
every assignment simply has a clone of the 'at that time' latest assignments and case structure.

a offline 'handbook' system is no different, pupils do not get new handbooks during a semester.