Implement modular dashboard system
Allows to easily create a controller with view to display a block on the dashboard.
For example to display user meta information, or a agenda

Added example:
Store the result of a case as json using the recently added (21e3d084c77d4379de7956cc3ceedd55cac0195c) user meta system.
Then retrieve that json and print it on the dashboard using a dedicated Controller & View